XML RPC for userpoints via integration with the services module.
This allows external applications and sites to retrieve and update
points on your site. Many things are possible with this, such as
a Drupal site acting as an award points or user reputation repository
for Joomla or Wordpress sites, or having user actions captured in
other systems (not necessarily web sites) updating the points on
your Drupal site. Use your imagination ...